# Plug memory leak caused by express-openapi-validate