--- a/dev/config-leaktest/gateway.config.yml
+++ b/dev/config-leaktest/gateway.config.yml
@@ -1,0 +1,227 @@
+http: # remove http section in production environment
+  port: 8080
+https:
+  port: 4444
+  tls:
+    default: # replace with real certificate in prod environment
+      key: "config/testServer.key"
+      cert: "config/testServer.crt"
+admin:
+  port: 9876
+  host: localhost
+apiEndpoints:
+  metrics-reporter:
+    paths:
+      - '/metrics'
+  api:
+    # all matching paths should be specified including any template
+    # params as used in the ooapi spec
+    paths:
+       - '/'
+       - '/academic-sessions'
+       - '/academic-sessions/:academicSessionId'
+       - '/academic-sessions/:academicSessionId/offerings'
+       - '/associations/:associationId'
+       - '/associations/external/me'
+       - '/buildings'
+       - '/buildings/:buildingId'
+       - '/buildings/:buildingId/rooms'
+       - '/components/:componentId'
+       - '/components/:componentId/offerings'
+       - '/courses'
+       - '/courses/:courseId'
+       - '/courses/:courseId/components'
+       - '/courses/:courseId/offerings'
+       - '/education-specifications'
+       - '/education-specifications/:educationSpecificationId'
+       - '/education-specifications/:educationSpecificationId/education-specifications'
+       - '/education-specifications/:educationSpecificationId/courses'
+       - '/education-specifications/:educationSpecificationId/programs'
+       - '/groups'
+       - '/groups/:groupId'
+       - '/groups/:groupId/persons'
+       - '/news-feeds'
+       - '/news-feeds/:newsFeedId'
+       - '/news-feeds/:newsFeedId/news-items'
+       - '/news-items/:newsItemId'
+       - '/offerings/:offeringId'
+       - '/offerings/:offeringId/associations'
+       - '/offerings/:offeringId/groups'
+       - '/organizations'
+       - '/organizations/:organizationId'
+       - '/organizations/:organizationId/components'
+       - '/organizations/:organizationId/courses'
+       - '/organizations/:organizationId/programs'
+       - '/organizations/:organizationId/offerings'
+       - '/organizations/:organizationId/groups'
+       - '/organizations/:organizationId/education-specifications'
+       - '/persons'
+       - '/persons/me'
+       - '/persons/:personId'
+       - '/persons/:personId/associations'
+       - '/persons/:personId/groups'
+       - '/programs'
+       - '/programs/:programId'
+       - '/programs/:programId/programs'
+       - '/programs/:programId/courses'
+       - '/programs/:programId/offerings'
+       - '/rooms'
+       - '/rooms/:roomId'
+  not_found: # fallback: needed to track non-matching requests
+    paths:
+       - '/*'
+serviceEndpoints:
+  'Test.Backend':
+    url: 'http://localhost:8082'
+    proxyOptions:
+      auth: fred:wilma
+policies:
+  - basic-auth
+  - cors
+  - expression
+  - key-auth
+  - lifecycle-logger
+  - proxy
+  - rate-limit
+  - openapi-validator
+  - gatekeeper
+  - request-transformer
+  - response-transformer
+  - aggregation
+  - metrics-collector
+  - metrics-reporter
+  - terminate
+pipelines:
+  metrics:
+    apiEndpoints:
+      - metrics-reporter
+    policies:
+      - metrics-reporter:
+          - action:
+              url: "/metrics"
+              labels:
+                testLabel: "test"
+                otherLabel: "foo"
+              defaultMetricsLabels:
+                def: "something"
+  test:
+    apiEndpoints:
+      - api
+    policies:
+      - metrics-collector:
+      - lifecycle-logger:
+      - gatekeeper:
+          - action:
+              apps:
+                fred:
+                  passwordSalt: 8b52795e90b598eb022647f17c93ac2b
+                  passwordHash: e4c9628c52aead0dcf10330a6864d8bcc78a5a4a463b274bee39cee4cfc0a574
+                barney:
+                  passwordSalt: 5e5b3fb149fdd06ba9d18edd178d77cb
+                  passwordHash: 19d767b82ebb294e3c2008b53a0bcc59140e688baded13eea1794099c869e89f
+                bubbles:
+                  passwordSalt: 5970ad7d7501916274cf114f68d2aed0
+                  passwordHash: 5e063ba6dcff4b7bc0234be7861dac8c4dd7db573f36755e0578b2e77a5cf6bf
+              acls:
+                - app: fred
+                  endpoints:
+                    - endpoint: 'Test.Backend'
+                      paths:
+                       - '/'
+                       - '/academic-sessions'
+                       - '/academic-sessions/:academicSessionId'
+                       - '/academic-sessions/:academicSessionId/offerings'
+                       - '/associations/:associationId'
+                       - '/associations/external/me'
+                       - '/buildings'
+                       - '/buildings/:buildingId'
+                       - '/buildings/:buildingId/rooms'
+                       - '/components/:componentId'
+                       - '/components/:componentId/offerings'
+                       - '/courses'
+                       - '/courses/:courseId'
+                       - '/courses/:courseId/components'
+                       - '/courses/:courseId/offerings'
+                       - '/education-specifications'
+                       - '/education-specifications/:educationSpecificationId'
+                       - '/education-specifications/:educationSpecificationId/education-specifications'
+                       - '/education-specifications/:educationSpecificationId/courses'
+                       - '/education-specifications/:educationSpecificationId/programs'
+                       - '/groups'
+                       - '/groups/:groupId'
+                       - '/groups/:groupId/persons'
+                       - '/news-feeds'
+                       - '/news-feeds/:newsFeedId'
+                       - '/news-feeds/:newsFeedId/news-items'
+                       - '/news-items/:newsItemId'
+                       - '/offerings/:offeringId'
+                       - '/offerings/:offeringId/associations'
+                       - '/offerings/:offeringId/groups'
+                       - '/organizations'
+                       - '/organizations/:organizationId'
+                       - '/organizations/:organizationId/components'
+                       - '/organizations/:organizationId/courses'
+                       - '/organizations/:organizationId/programs'
+                       - '/organizations/:organizationId/offerings'
+                       - '/organizations/:organizationId/groups'
+                       - '/organizations/:organizationId/education-specifications'
+                       - '/persons'
+                       - '/persons/me'
+                       - '/persons/:personId'
+                       - '/persons/:personId/associations'
+                       - '/persons/:personId/groups'
+                       - '/programs'
+                       - '/programs/:programId'
+                       - '/programs/:programId/programs'
+                       - '/programs/:programId/courses'
+                       - '/programs/:programId/offerings'
+                       - '/rooms'
+                       - '/rooms/:roomId'
+
+      - request-transformer:
+          - action:
+              headers:
+                add:
+                  'Accept': '"application/json"'
+      - response-transformer:
+          - action:
+              headers:
+                add:
+                  # note: literal header values should be quoted twice
+                  # see https://www.express-gateway.io/docs/policies/response-transformer/
+                  'X-XSS-Protection': "'1; mode-block'"
+                  'Strict-Transport-Security': "'max-age=31536000; includeSubDomains; preload'"
+                  'X-Frame-Options': "'SAMEORIGIN'"
+                  'X-Content-Type-Options': "'nosniff'"
+                  # note: double escaping because of YAML parsing
+                  'Content-Security-Policy': "'default-src \\'self\\''"
+                  'Access-Control-Allow-Origin': "'*.surf.nl'"
+                  'Referrer-Policy': "'no-referrer-when-downgrade'"
+
+      - openapi-validator:
+          - action:
+              apiSpec: 'ooapiv5.json'
+              validateRequests: true
+              validateResponses: true
+
+      - aggregation:
+          - action:
+              noEnvelopIfAnyHeaders:
+                'X-Validate-Response': 'true'
+              keepRequestHeaders:
+                - 'accept-language'
+                - 'accept'
+                - 'x-validate-response' # needed for resp validation
+              keepResponseHeaders:
+                - 'content-type'
+                - 'content-length'
+  not_found:
+    apiEndpoints:
+      - not_found
+    policies:
+      - metrics-collector:
+      - lifecycle-logger:
+      - terminate:
+          - action:
+              statusCode: 404
+              message: "Not found"