--- a/dev/config-leaktest/models/applications.json
+++ b/dev/config-leaktest/models/applications.json
@@ -1,0 +1,16 @@
+{
+  "$id": "http://express-gateway.io/models/applications.json",
+  "type": "object",
+  "properties": {
+    "name": {
+      "type": "string"
+    },
+    "redirectUri": {
+      "type": "string",
+      "format": "uri"
+    }
+  },
+  "required": [
+    "name"
+  ]
+}