--- a/dev/config-leaktest/system.config.yml.test
+++ b/dev/config-leaktest/system.config.yml.test
@@ -1,0 +1,26 @@
+db:
+  redis:
+    namespace: EG
+    host: '${REDIS_HOST}'
+    port: '${REDIS_PORT}'
+
+#plugins:
+  # express-gateway-plugin-example:
+  #   param1: 'param from system.config'
+
+crypto:
+  cipherKey: sensitiveKey
+  algorithm: aes256
+  saltRounds: 10
+
+# OAuth2 Settings
+session:
+  secret: keyboard cat
+  resave: false
+  saveUninitialized: false
+accessTokens:
+  timeToExpiry: 7200000
+refreshTokens:
+  timeToExpiry: 7200000
+authorizationCodes:
+  timeToExpiry: 300000