The `express-openapi-validate` library used to validate requests and responses leaks memory when a validator object is reused.  This changeset creates a new validator on every request which is costly.  Below are some basic test results (using Node v18.13 on an Intel(R) Core(TM) i7-10710U with plenty of RAM).

Long store short: only request validation will cause a performance reduction to 59% of the original, with response validation 42%.

# Request validation

## before

```
Concurrency Level:      1
Time taken for tests:   31.855 seconds
Complete requests:      5000
Failed requests:        0
Total transferred:      16215000 bytes
HTML transferred:       13790000 bytes
Requests per second:    156.96 [#/sec] (mean)
Time per request:       6.371 [ms] (mean)
Time per request:       6.371 [ms] (mean, across all concurrent requests)
Transfer rate:          497.10 [Kbytes/sec] received
```

## after

```
Concurrency Level:      1
Time taken for tests:   53.671 seconds
Complete requests:      5000
Failed requests:        0
Total transferred:      16215000 bytes
HTML transferred:       13790000 bytes
Requests per second:    93.16 [#/sec] (mean)
Time per request:       10.734 [ms] (mean)
Time per request:       10.734 [ms] (mean, across all concurrent requests)
Transfer rate:          295.04 [Kbytes/sec] received
```

Performance hit: 59% of original

# Request and response validation

## before

```
Concurrency Level:      1
Time taken for tests:   36.109 seconds
Complete requests:      5000
Failed requests:        0
Total transferred:      16215000 bytes
HTML transferred:       13790000 bytes
Requests per second:    138.47 [#/sec] (mean)
Time per request:       7.222 [ms] (mean)
Time per request:       7.222 [ms] (mean, across all concurrent requests)
Transfer rate:          438.53 [Kbytes/sec] received
```

## after

```
Concurrency Level:      1
Time taken for tests:   85.425 seconds
Complete requests:      5000
Failed requests:        0
Total transferred:      16215000 bytes
HTML transferred:       13790000 bytes
Requests per second:    58.53 [#/sec] (mean)
Time per request:       17.085 [ms] (mean)
Time per request:       17.085 [ms] (mean, across all concurrent requests)
Transfer rate:          185.37 [Kbytes/sec] received
```

Performance hit: 42% of original.